refactor: ViewController에 MenuCollectionViewController 통합
